Combine Python tests for Unit
