+++ tests/unit_test.py
@@ -16,11 +16,9 @@ def test_default_unit():
     assert u == sc.units.dimensionless
 
 
-def test_unit_repr():
-    assert (repr(sc.units.angstrom) == "\u212B") or (repr(sc.units.angstrom)
-                                                     == "\u00C5")
-    assert repr(sc.units.m) == "m"
-    assert repr(sc.units.us) == "µs"
+def test_can_construct_unit_from_string():
+    assert sc.Unit('m') == sc.units.m
+    assert sc.Unit('meV') == sc.units.meV
 
 
 def test_unit_from_unicode():
@@ -31,3 +29,42 @@ def test_unit_from_unicode():
 def test_unit_equal_unicode():
     assert sc.Unit('angstrom') == 'Å'
     assert sc.Unit('us') == 'µs'
+
+
+def test_constructor_raises_with_bad_input():
+    with pytest.raises(sc.UnitError):
+        sc.Unit('abcdef')  # does not parse
+    with pytest.raises(TypeError):
+        sc.Unit(5)  # neither str nor Unit
+
+
+def test_unit_repr():
+    assert repr(sc.units.m) == 'm'
+    assert repr(sc.units.us) == 'µs'
+
+
+@pytest.mark.parametrize('u', (sc.units.angstrom, sc.Unit('angstrom')))
+def test_angstrom_repr(u):
+    assert repr(u) == '\u212B' or repr(u) == '\u00C5'
+
+
+def test_unit_property_from_str():
+    var = sc.scalar(1)
+    var.unit = 'm'
+    assert var.unit == sc.Unit('m')
+
+
+def test_unit_property_from_unit():
+    var = sc.scalar(1)
+    var.unit = sc.Unit('s')
+    assert var.unit == sc.Unit('s')
+
+
+def test_explicit_default_unit_for_string_gives_none():
+    var = sc.scalar('abcdef', unit=sc.units.default_unit)
+    assert var.unit is None
+
+
+def test_default_unit_for_string_is_none():
+    var = sc.scalar('abcdef')
+    assert var.unit is None
